+++ web/src/app/dashboard/components/FeedbackPanel.tsx
@@ -12,7 +12,7 @@ import {
   useFeedbacks,
   useResolveFeedback,
   useUnresolveFeedback,
-} from "~/hooks/useDashboard";
+} from "~/hooks/dashboard/useFeedback";
 import { useInfiniteScroll } from "~/hooks/useInfiniteScroll";
 import { confirmDialog } from "~/lib/dialog";
 import { formatDateTime } from "~/lib/format";
