+++ web/src/app/dashboard/components/knowledge/FileUpload.tsx
@@ -1,6 +1,6 @@
 import { ArrowUpTrayIcon } from "@heroicons/react/24/outline";
 import { useCallback, useState } from "react";
-import { useConvertFile, useUploadFile } from "~/hooks/useDashboard";
+import { useConvertFile, useUploadFile } from "~/hooks/dashboard/useKnowledge";
 import { ConvertDialog } from "./ConvertDialog";
 
 type Props = {
