+++ web/src/hooks/useDashboard.test.ts
@@ -7,12 +7,7 @@ import { renderHookWithQuery } from "~/test/query";
 import {
   useBroadcasts,
   useCreateBroadcast,
-  useFeedbackDetail,
-  useKnowledgeFile,
-  useKnowledgeFiles,
   usePollResults,
-  useSyncKnowledge,
-  useUnifiedFiles,
 } from "./useDashboard";
 
 const API = "http://localhost:8787";
@@ -27,57 +22,9 @@ afterEach(() => {
 });
 
 describe("条件付き enabled 系 query", () => {
-  it("useKnowledgeFile: key=null なら fetchStatus は idle", () => {
-    const { result } = renderHookWithQuery(() => useKnowledgeFile(null));
-    expect(result.current.fetchStatus).toBe("idle");
-  });
-
-  it("useFeedbackDetail: id=null なら fetchStatus は idle", () => {
-    const { result } = renderHookWithQuery(() => useFeedbackDetail(null));
-    expect(result.current.fetchStatus).toBe("idle");
-  });
-
   it("usePollResults: id=null なら fetchStatus は idle", () => {
     const { result } = renderHookWithQuery(() => usePollResults(null));
     expect(result.current.fetchStatus).toBe("idle");
-  });
-
-  it("useKnowledgeFile: key 指定時に fetch される", async () => {
-    server.use(
-      http.get(`${API}/admin/knowledge/files/doc.md`, () =>
-        HttpResponse.json({ content: "# title" }),
-      ),
-    );
-
-    const { result } = renderHookWithQuery(() => useKnowledgeFile("doc.md"));
-
-    await waitFor(() => expect(result.current.isSuccess).toBe(true));
-    expect(result.current.data?.content).toBe("# title");
-  });
-});
-
-describe("単純 query", () => {
-  it("useKnowledgeFiles: 200 を返したら data に反映", async () => {
-    server.use(
-      http.get(`${API}/admin/knowledge/files`, () =>
-        HttpResponse.json({ files: [] }),
-      ),
-    );
-
-    const { result } = renderHookWithQuery(() => useKnowledgeFiles());
-    await waitFor(() => expect(result.current.isSuccess).toBe(true));
-    expect(result.current.data?.files).toEqual([]);
-  });
-
-  it("useUnifiedFiles", async () => {
-    server.use(
-      http.get(`${API}/admin/knowledge/unified`, () =>
-        HttpResponse.json({ files: [] }),
-      ),
-    );
-
-    const { result } = renderHookWithQuery(() => useUnifiedFiles());
-    await waitFor(() => expect(result.current.isSuccess).toBe(true));
   });
 });
 
@@ -101,21 +48,6 @@ describe("infinite query 系", () => {
 });
 
 describe("mutation 系", () => {
-  it("useSyncKnowledge: 単純 mutation", async () => {
-    server.use(
-      http.post(`${API}/admin/knowledge/sync`, () =>
-        HttpResponse.json({ message: "ok" }),
-      ),
-    );
-
-    const { result } = renderHookWithQuery(() => useSyncKnowledge());
-
-    await act(async () => {
-      await result.current.mutateAsync();
-    });
-    await waitFor(() => expect(result.current.isSuccess).toBe(true));
-  });
-
   it("useCreateBroadcast: 成功後に broadcasts を invalidate", async () => {
     server.use(
       http.post(`${API}/admin/broadcast`, () =>
