+++ web/src/hooks/useDashboard.ts
@@ -4,181 +4,11 @@ import {
   useQuery,
   useQueryClient,
 } from "@tanstack/react-query";
-import {
-  broadcastRepository,
-  feedbackRepository,
-  knowledgeRepository,
-  pollRepository,
-} from "~/lib/api/repository";
+import { broadcastRepository, pollRepository } from "~/lib/api/repository";
 import type { PollStatus } from "~/types";
 import { dashboardKeys } from "./dashboard/keys";
 
 export { dashboardKeys };
-
-// ナレッジファイル関連 hooks
-export const useSyncKnowledge = () =>
-  useMutation({
-    mutationFn: knowledgeRepository.syncKnowledge,
-  });
-
-export const useDeleteKnowledge = () =>
-  useMutation({
-    mutationFn: knowledgeRepository.deleteAllKnowledge,
-  });
-
-export const useKnowledgeFiles = () =>
-  useQuery({
-    queryKey: dashboardKeys.knowledgeFiles,
-    queryFn: knowledgeRepository.fetchFiles,
-  });
-
-export const useKnowledgeFile = (key: string | null) =>
-  useQuery({
-    queryKey: dashboardKeys.knowledgeFile(key ?? ""),
-    queryFn: () => {
-      if (!key) throw new Error("Key is required");
-      return knowledgeRepository.fetchFileContent(key);
-    },
-    enabled: !!key,
-  });
-
-export const useSaveFile = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: ({ key, content }: { key: string; content: string }) =>
-      knowledgeRepository.saveFile(key, content),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: dashboardKeys.knowledgeFiles });
-      queryClient.invalidateQueries({
-        queryKey: dashboardKeys.knowledgeUnifiedFiles,
-      });
-    },
-  });
-};
-
-export const useDeleteFile = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: knowledgeRepository.deleteFile,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: dashboardKeys.knowledgeFiles });
-      queryClient.invalidateQueries({
-        queryKey: dashboardKeys.knowledgeUnifiedFiles,
-      });
-    },
-  });
-};
-
-export const useUploadFile = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: ({ file, filename }: { file: File; filename?: string }) =>
-      knowledgeRepository.uploadFile(file, filename),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: dashboardKeys.knowledgeFiles });
-      queryClient.invalidateQueries({
-        queryKey: dashboardKeys.knowledgeUnifiedFiles,
-      });
-    },
-  });
-};
-
-export const useConvertFile = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: ({ file, filename }: { file: File; filename: string }) =>
-      knowledgeRepository.convertFile(file, filename),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: dashboardKeys.knowledgeFiles });
-      queryClient.invalidateQueries({
-        queryKey: dashboardKeys.knowledgeUnifiedFiles,
-      });
-    },
-  });
-};
-
-// 統合ファイル一覧
-export const useUnifiedFiles = () =>
-  useQuery({
-    queryKey: dashboardKeys.knowledgeUnifiedFiles,
-    queryFn: knowledgeRepository.fetchUnifiedFiles,
-  });
-
-// 元ファイルからMarkdownを再生成
-export const useReconvertFile = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: ({
-      originalKey,
-      filename,
-    }: {
-      originalKey: string;
-      filename: string;
-    }) => knowledgeRepository.reconvertFile(originalKey, filename),
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: dashboardKeys.knowledgeUnifiedFiles,
-      });
-    },
-  });
-};
-
-// フィードバック関連 hooks
-export const useFeedbacks = (
-  limit = 30,
-  options?: { rating?: "good" | "bad" | "idea" },
-) =>
-  useInfiniteQuery({
-    queryKey: [...dashboardKeys.feedbacks, limit, options?.rating],
-    queryFn: ({ pageParam }) =>
-      feedbackRepository.fetchFeedbacks({
-        limit,
-        cursor: pageParam,
-        rating: options?.rating,
-      }),
-    initialPageParam: undefined as string | undefined,
-    getNextPageParam: (lastPage) => lastPage.nextCursor ?? undefined,
-  });
-
-export const useFeedbackDetail = (id: string | null) =>
-  useQuery({
-    queryKey: dashboardKeys.feedbackDetail(id ?? ""),
-    queryFn: () => {
-      if (!id) throw new Error("ID is required");
-      return feedbackRepository.fetchFeedbackById(id);
-    },
-    enabled: !!id,
-  });
-
-export const useDeleteFeedbacks = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: feedbackRepository.deleteAllFeedbacks,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: dashboardKeys.feedbacks });
-    },
-  });
-};
-
-export const useResolveFeedback = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: feedbackRepository.resolveFeedback,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: dashboardKeys.feedbacks });
-    },
-  });
-};
-
-export const useUnresolveFeedback = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: feedbackRepository.unresolveFeedback,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: dashboardKeys.feedbacks });
-    },
-  });
-};
 
 // 配信メッセージ関連 hooks
 export const useBroadcasts = (
